ex00

+++ d02/d02_ex00/d02_ex00/Models/ExchangeRate.cs
@@ -10,9 +10,9 @@ namespace d02_ex00.Models
     {
         public string FromCurrency { get; }
         public string ToCurrency { get; }
-        public decimal Rate { get; }
+        public double Rate { get; }
 
-        public ExchangeRate(string fromCurrency, string toCurrency, decimal rate)
+        public ExchangeRate(string fromCurrency, string toCurrency, double rate)
         {
             FromCurrency = fromCurrency;
             ToCurrency = toCurrency;
